+++ .termux_tui_music_config.json
@@ -1,0 +1,7 @@
+{
+  "music_dirs": [
+    "/storage/C4C9-469E/Audio/"
+  ],
+  "music_mode": "sequential",
+  "music_stop_on_close": false
+}
